feat: form ui modified
