+++ frontend/src/components/Form/FormView.js
@@ -1,84 +1,97 @@
-import { Button, Box, Form, FormField, TextInput, Select } from 'grommet';
-import ImageUploadView from './ImageUploadView';
-import useForm from './useForm';
-import formHandler from './formHandler';
+import { useState, forwardRef } from 'react';
+import {
+  AppBar,
+  Toolbar,
+  Fab,
+  Dialog,
+  DialogTitle,
+  DialogContent,
+  DialogContentText,
+  DialogActions,
+  Button,
+  makeStyles,
+  Slide,
+  IconButton,
+  Typography,
+} from '@material-ui/core';
+import AddIcon from '@material-ui/icons/Add';
+import CloseIcon from '@material-ui/icons/Close';
+
+import Form from './Form';
+
+const useStyles = makeStyles((theme) => ({
+  extendedIcon: {
+    marginRight: theme.spacing(1),
+  },
+  fab: {
+    position: 'fixed',
+    bottom: theme.spacing(2),
+    right: theme.spacing(2),
+  },
+  appBar: {
+    position: 'relative',
+  },
+  title: {
+    marginLeft: theme.spacing(2),
+    flex: 1,
+  },
+}));
+
+const Transition = forwardRef(function Transition(props, ref) {
+  return <Slide direction='up' ref={ref} {...props} />;
+});
 
 function FormView() {
-  const [formState, inputHandler] = useForm(
-    {
-      title: {
-        value: '',
-        isValid: false,
-      },
-      description: {
-        value: '',
-        isValid: false,
-      },
-      address: {
-        value: '',
-        isValid: false,
-      },
-      coordinates: {
-        value: '',
-        isValid: false,
-      },
-      image: {
-        value: null,
-        isValid: false,
-      },
-    },
-    false
-  );
+  const classes = useStyles();
+
+  const [modalIsOpen, setIsOpen] = useState(false);
 
   return (
-    <Box flex justify='around' background='dark-2'>
-      <Form
-        value={{}}
-        onReset={() => {}}
-        onSubmit={(evt) => formHandler(evt, formState)}
+    <>
+      <Fab
+        className={classes.fab}
+        variant='extended'
+        onClick={() => {
+          setIsOpen((prev) => !prev);
+        }}
       >
-        <FormField name='name' htmlfor='text-input-id' label='Name'>
-          <TextInput id='text-input-id' name='name' />
-        </FormField>
-        <FormField name='phone' htmlfor='text-input-id' label='Phone No.'>
-          <TextInput id='text-input-id2' name='phone' />
-        </FormField>
-        <FormField
-          name='description'
-          htmlfor='text-input-id'
-          label='Item Description'
-        >
-          <TextInput id='text-input-id3' name='description' />
-        </FormField>
-        <FormField name='address' htmlfor='text-input-id' label='Address'>
-          <TextInput id='text-input-id4' name='address' />
-        </FormField>
-        <FormField name='country' htmlfor='text-input-id' label='Country'>
-          <TextInput id='text-input-id5' name='country' />
-        </FormField>
-        <FormField name='city' htmlfor='text-input-id' label='City'>
-          <TextInput id='text-input-id6' name='city' />
-        </FormField>
-        <FormField name='category' htmlfor='text-input-id' label='Category'>
-          <Select
-            options={['household', 'electrical', 'metal']}
-            name='category'
-          />
-        </FormField>
-        <FormField name='image' htmlfor='text-input-id' label='Image'>
-          <ImageUploadView
-            id='image'
-            name='image'
-            onInput={inputHandler}
-            errorText='Please provide an image.'
-          />
-        </FormField>
-        <Box direction='row' gap='medium'>
-          <Button type='submit' primary label='Submit' />
-          <Button type='reset' label='Reset' />
-        </Box>
-      </Form>
-    </Box>
+        <AddIcon className={classes.extendedIcon} />
+        Create New Post
+      </Fab>
+      <Dialog
+        fullScreen
+        open={modalIsOpen}
+        onClose={() => setIsOpen(false)}
+        contentLabel='Add your details here'
+        TransitionComponent={Transition}
+      >
+        <AppBar className={classes.appBar}>
+          <Toolbar>
+            <IconButton
+              edge='start'
+              color='inherit'
+              onClick={() => setIsOpen(false)}
+              aria-label='close'
+            >
+              <CloseIcon />
+            </IconButton>
+            <Typography variant='h6' className={classes.title}>
+              Create New Post
+            </Typography>
+            <Button autoFocus color='inherit' onClick={() => setIsOpen(false)}>
+              Save
+            </Button>
+          </Toolbar>
+        </AppBar>
+        <DialogContent>
+          <DialogContentText>
+            To add your personalised ad here, submit the below form. Sell your
+            old stuff easily
+          </DialogContentText>
+          <Form />
+        </DialogContent>
+      </Dialog>
+    </>
   );
 }
